pyproject.toml
[project]
name = "python-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Fathur Rohman",email = "hi.fathur.rohman@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "ipython (>=9.0.2,<10.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "random-english-word-generator (>=1.0.1,<2.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.0"
factory-boy = "^3.3.3"
pre-commit = "^4.2.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
addopts = [
    "--reuse-db",
    "--nomigrations"
]
